+++ src/redforge/reasoning/__init__.py
@@ -1,0 +1,33 @@
+from .contracts import Goal, ReasoningDecision
+from .exceptions import ReasoningError, GoalError, StrategyError
+from .goal_manager import GoalManager
+from .task_decomposer import TaskDecomposer
+from .strategy_selector import StrategySelector
+from .world_state import WorldState
+from .self_evaluator import SelfEvaluator
+from .reflection import SelfReflection
+from .replanner import Replanner
+from .failure_handler import FailureHandler
+from .state_machine import ReasoningState, ReasoningStateMachine
+from .reasoner import Reasoner
+from .engine import ReasoningEngine
+
+__all__ = [
+    "Goal",
+    "ReasoningDecision",
+    "ReasoningError",
+    "GoalError",
+    "StrategyError",
+    "GoalManager",
+    "TaskDecomposer",
+    "StrategySelector",
+    "WorldState",
+    "SelfEvaluator",
+    "SelfReflection",
+    "Replanner",
+    "FailureHandler",
+    "ReasoningState",
+    "ReasoningStateMachine",
+    "Reasoner",
+    "ReasoningEngine"
+]